+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # ts2workflows changelog
 
-## unrelease
+## unreleased
 
-- expressions as properties in assignments: `a[i + 4] = 1` (regression in version 0.9.0)
+- Accept binary expressions as properties in assignments: `a[i + 4] = 1` (regression since 0.9.0)
+- Evaluate side-effects only once on the left-hand side in compound assignments: `x[f()] += 1`
 
 ## Version 0.9.0 - 2025-04-23